+++ Sources/csuper-gui/new_file_assistant.c
@@ -501,7 +501,8 @@ G_MODULE_EXPORT void endAssistantNewCsu(GtkWidget *widget, gpointer data)
     user_data->ptr_csu_struct_tmp = NULL;
 
     updateMainWindow(user_data);
-    writeCsuFile(user_data->csu_filename,user_data->ptr_csu_struct);
+    if (writeCsuFile(user_data->csu_filename,user_data->ptr_csu_struct) == MY_FALSE)
+        saveFileError(user_data);
     deleteAllLastCsuStruct(user_data);
     addLastCsuStruct(user_data);